requirements.txt
scipy==1.0.0
numpy==1.14.0
Pillow==5.0.0
